fixing compilation error

+++ 0x0B-malloc_free/0-create_array.c
@@ -25,7 +25,7 @@ char *create_array(unsigned int size, char c)
 	}
 	else
 	{
-		for (i = 0; i < size; i++)
+		for (i = 0; i < (int) size; i++)
 		{
 			*(str + i) = c;
 		}
